requirements.txt
flask
faiss-cpu
numpy
sentence-transformers